requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
pytest>=7.0
requests>=2.31.0
